backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
feedparser==6.0.11
httpx==0.27.2
telethon==1.37.0
python-dotenv==1.0.1
python-multipart==0.0.12
websockets==13.1
openai==1.82.0
